Corrected installation failure for requirements.txt
if not on windows the `output` variable was set to "/dev/null" which meant that /dev/null would be seen as a requirements.txt file for pip. I added the `> ` to pipe stdout to /dev/null instead of specifying it as a requirements.txt. 
I also changed the command from `py` to `python3` because I've never seen the command `py` for python before. 
I even installed a Window$ 10 VM to test if `python3` would work as a  command on Window$. It does.